backend/requirements/common.txt
blinker==1.4
coverage==4.0.2
Flask==0.12.5
Flask-Script==2.0.5
Flask-Testing==0.4.2
funcsigs==0.4
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
MarkupSafe==0.23
mock==1.3.0
nose==1.3.7
pbr==1.8.1
pycparser==2.14
pytz==2015.7
requests==2.23.0
selenium==2.48.0
six==1.10.0
traceback2==1.4.0
twilio==6.9.0
unittest2==1.1.0
Werkzeug==0.11.2
python-dotenv==0.13.0
